fix(chat): Prevent input clearing when clicking chat buttons
The chat box is no longer cleared when clicking buttons

Previously, we had a bug where if either of the buttons in the agent chat was clicked, the ChatTextArea would get cleared. Now, the text box will only get cleared if the message is sent as part of the response.

Moved input clearing logic to only execute when a message is successfully sent,
preventing the input field from clearing prematurely when clicking "yes" or "no"
buttons in the chat interface. This improves user experience by retaining the
input content if an action fails or requires further interaction

Fixes #1520